+ converted error handling to 16 bit signed to integrate RadioLib errors into the handler

+++ include/cmds/bitr.hpp
@@ -8,6 +8,6 @@
 #define BITR_MIN 1.2
 #define BITR_MAX 300.0
 
-uint8_t bitr(float bitrate, RFM95 &radio);
+int16_t bitr(float bitrate, RFM95 &radio);
 
 #endif
