+++ src/cmds/echo.cpp
@@ -1,6 +1,6 @@
 #include "cmds/echo.hpp"
 
-uint8_t echo()
+int16_t echo()
 {
     Serial.println(F("Hello, world!"));
     return RUN_SUCCESS;
