+++ src/cmds/help.cpp
@@ -1,6 +1,6 @@
 #include "cmds/help.hpp"
 
-uint8_t help()
+int16_t help()
 {
     // TODO: Implement Help command
     return NOT_YET_IMPLEMENTED;
